requirements/requirements.txt
Django==3.0.6
django-crispy-forms
django-filter
django-channels
